jwt authorization code is completed.

+++ controller/api/v1/posts_api.js
@@ -23,8 +23,8 @@ module.exports.destroy = async function(req, res){
     try {
         let post=await Post.findById(req.params.id);
 
-        // if(post.user == req.user.id)
-        // {
+        if(post.user == req.user.id)
+        {
             post.remove();
 
             await Comment.deleteMany({post: req.params.id});
@@ -33,7 +33,13 @@ module.exports.destroy = async function(req, res){
                 return res.json(200,{
                    message: "Post deleted successfully."
                 });
-        // }
+        }
+        else
+        {
+            return res.json(401, {
+                message: "You can not delete this post!"
+            });
+        }
     } catch (error) {
         return res.json(500, {
             message : "Internal srver error"
